+++ wstx/trunk/src/java/com/ctc/wstx/io/WstxInputSource.java
@@ -106,7 +106,7 @@ public abstract class WstxInputSource
      */
     protected abstract WstxInputLocation getLocation();
 
-    public abstract WstxInputLocation getLocation(int total, int row, int col);
+    public abstract WstxInputLocation getLocation(long total, int row, int col);
 
     /*
     //////////////////////////////////////////////////////////
